fix: Avoid signing out users when the token refresh failed due to connectivity issues on startup
